# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with BookStore Application",
        "description": "Integrate Swagger in Laravel  BookStoreapplication",
        "contact": {
            "email": "chandanmohanty11111@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/forgotpassword": {
            "post": {
                "summary": "forgotpassword",
                "description": "Send Mail to the respectice mail id for forget password link",
                "operationId": "186a73907a51e79b61ee87564453da94",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "we have mailed your password reset link to respective E-mail"
                    },
                    "404": {
                        "description": "we can not find a user with that email address"
                    }
                }
            }
        },
        "/api/auth/resetpassword": {
            "post": {
                "summary": "resetpassword",
                "description": "reset your password",
                "operationId": "b3e651e8d03d5635958aef7d32e69961",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "new_password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "new_password": {
                                        "type": "password"
                                    },
                                    "confirm_password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Password reset successfull!"
                    },
                    "400": {
                        "description": "we can not find the user with that e-mail address"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "summary": "register",
                "description": "register the user for login",
                "operationId": "47d5ee6b3638073a988b692b1a5902cf",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "role",
                                    "firstname",
                                    "lastname",
                                    "phone_no",
                                    "email",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "phone_no": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "confirm_password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully registered"
                    },
                    "401": {
                        "description": "The email has already been taken"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "login",
                "description": " login ",
                "operationId": "224ca9cc21e92b4e3291419882455b09",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successfull"
                    },
                    "401": {
                        "description": "we can not find the user with that e-mail address You need to register first"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "logout",
                "description": " logout the user or admin ",
                "operationId": "53d83c37c7970aebebdc61ca4971d22e",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "User successfully signed out"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}